+++ Back End/Controllers/ReviewController.cs
@@ -8,6 +8,7 @@ using Back_End.DTOs.Review;
 using Back_End.Mappers;
 using Back_End.Models;
 using Back_End.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Back_End.Controllers
@@ -24,38 +25,62 @@ namespace Back_End.Controllers
         }
 
         [HttpGet]
+        [Authorize]
+
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var reviews = await _reviewService.GetAllAsync();
             return Ok(reviews);
-        }       
+        }
 
         [HttpGet("{id:int}")]
+        [Authorize]
+
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var review = await _reviewService.GetByIdAsync(id);
             return Ok(review);
         }
 
         [HttpPost]
         [Route("{itemId:int}")]
-        public async Task<IActionResult> Create([FromRoute] int itemId,   CreateReviewRequestDto createReviewRequestDto)
+        [Authorize]
+
+        public async Task<IActionResult> Create([FromRoute] int itemId, CreateReviewRequestDto createReviewRequestDto)
         {
-            ReviewModel review = await _reviewService.CreateAsync(itemId,createReviewRequestDto);
-            
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            ReviewModel review = await _reviewService.CreateAsync(itemId, createReviewRequestDto);
+
             return CreatedAtAction(nameof(GetById), new { id = review.Id }, review.ToReviewResponseDto());
         }
 
         [HttpPut("{id:int}")]
-        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateReviewRequestDto updateReviewRequestDto) 
+        [Authorize]
+
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateReviewRequestDto updateReviewRequestDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             await _reviewService.UpdateAsync(id, updateReviewRequestDto);
             return NoContent();
         }
 
         [HttpDelete("{id:int}")]
+        [Authorize]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             await _reviewService.DeleteAsync(id);
             return NoContent();
         }
